ADD:「values.go」を作成し、コンテキストの値の取得・設定を移管。
